Archipelago: Pet Shop / Task Display Options.
Generates and functions, but haven't tested it further than checking if it works as expected in game.

might want to add another choice that just tells you who it's for and the class of item. `Your Progression item" or "Someone's Trap" and like that, or even just who it's for.